refactor(extension-common): 広告リンク検証のメッセージを専用メッセンジャーへ分ける
CredentialsProtocolMap が fetchCredentials / verifyIntegrity に加えて
adClicked / getVerificationResult を持っており、4 つのうち 2 つは広告リンク検証
の関心だった。検知の実装を link-verification へ移しても、メッセージ定義が
credentials に残っていると、移した側が credentials のメッセンジャーを import し
返す形になり境界が半分しか引かれない。

link-verification/events.ts に linkVerificationMessenger を作り、この 2 つを
移す。fetchVerificationResult も credentials/messaging.ts から
link-verification/messaging.ts へ移した。

LinkVerificationResult も credentials/types.ts から link-verification/types.ts
へ移す。分離後は credentials 側でこの型を使う箇所がなくなるため。パッケージの
バレルからは引き続き export されるので、inspector 側の import は変わらない。
