+++ packages/extension-common/src/link-verification/events.ts
@@ -1,0 +1,15 @@
+import { defineExtensionMessaging } from "@webext-core/messaging";
+import type { LinkVerificationResult } from "./types";
+
+type LinkVerificationProtocolMap = {
+  adClicked(message: {
+    targetopid: string;
+    sourceOrgName?: string;
+    expectedOrgName?: string;
+    isNewTab?: boolean;
+  }): void;
+  getVerificationResult(tabId: number): LinkVerificationResult;
+};
+
+export const linkVerificationMessenger =
+  defineExtensionMessaging<LinkVerificationProtocolMap>();
